Tests/LibWeb/Text/expected/shared-array-buffer-basic.txt

byteLength: 8
view[0]: 42
Atomics.load view[1]: 7
